back-rpod-setup/api/requirements.txt
fastapi
uvicorn
requests
requests-unixsocket
python-dotenv
pyyaml
python-multipart